Refactor: domain과 controller 분리
