--- src/main/java/com/unsolvedwa/unsolvedwa/BaseTimeEntity.java
@@ -1,23 +1,0 @@
-package com.unsolvedwa.unsolvedwa;
-
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.Column;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-
-@Getter
-@MappedSuperclass //공통맵핑정보
-@EntityListeners(AuditingEntityListener.class) //entity를감시?
-public abstract class BaseTimeEntity {
-    @CreatedDate
-    @Column(updatable = false, nullable = false)
-    private LocalDateTime createAt;
-
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
-}
