+++ src/main/java/com/unsolvedwa/unsolvedwa/domain/problem/ProblemRepository.java
@@ -1,0 +1,7 @@
+package com.unsolvedwa.unsolvedwa.domain.problem;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProblemRepository extends JpaRepository<Problem, Long>, ProblemRepositoryCustom {
+
+}
