+++ src/main/java/com/unsolvedwa/unsolvedwa/domain/problem/ProblemRepositoryCustom.java
@@ -1,0 +1,6 @@
+package com.unsolvedwa.unsolvedwa.domain.problem;
+
+import java.util.List;
+
+public interface ProblemRepositoryCustom {
+}
