+++ src/main/java/com/unsolvedwa/unsolvedwa/domain/problem_team/ProblemTeamRepository.java
@@ -1,0 +1,6 @@
+package com.unsolvedwa.unsolvedwa.domain.problem_team;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProblemTeamRepository extends JpaRepository<ProblemTeam, Long>, ProblemTeamRepositoryCustom {
+}
