+++ src/main/java/com/unsolvedwa/unsolvedwa/domain/problem_team/ProblemTeamRepositoryCustom.java
@@ -1,0 +1,5 @@
+package com.unsolvedwa.unsolvedwa.domain.problem_team;
+
+public interface ProblemTeamRepositoryCustom {
+    ProblemTeam[] solvingProblem(Long user_id, Long boj_id);
+}
