+++ src/main/java/com/unsolvedwa/unsolvedwa/domain/team/TeamRepository.java
@@ -1,0 +1,7 @@
+package com.unsolvedwa.unsolvedwa.domain.team;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TeamRepository extends JpaRepository<Team, Long>, TeamRepositoryCustom {
+
+}
