+++ src/main/java/com/unsolvedwa/unsolvedwa/domain/team/TeamRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.unsolvedwa.unsolvedwa.domain.team;
+
+public interface TeamRepositoryCustom {
+}
