+++ src/main/java/com/unsolvedwa/unsolvedwa/domain/team/TeamRepositoryImpl.java
@@ -1,0 +1,11 @@
+package com.unsolvedwa.unsolvedwa.domain.team;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class TeamRepositoryImpl implements TeamRepositoryCustom {
+
+    private JPAQueryFactory queryFactory;
+
+}
