+++ src/main/java/com/unsolvedwa/unsolvedwa/domain/user/UserRepository.java
@@ -1,0 +1,6 @@
+package com.unsolvedwa.unsolvedwa.domain.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long>, UserRepositoryCustom{
+}
