+++ src/main/java/com/unsolvedwa/unsolvedwa/domain/user/UserRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.unsolvedwa.unsolvedwa.domain.user;
+
+public interface UserRepositoryCustom {
+}
