+++ src/main/java/com/unsolvedwa/unsolvedwa/domain/user/UserRepositoryImpl.java
@@ -1,0 +1,7 @@
+package com.unsolvedwa.unsolvedwa.domain.user;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class UserRepositoryImpl implements UserRepositoryCustom{
+}
