+++ src/main/java/com/unsolvedwa/unsolvedwa/domain/user/UserService.java
@@ -1,0 +1,17 @@
+package com.unsolvedwa.unsolvedwa.domain.user;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service("userService")
+@Transactional
+@RequiredArgsConstructor
+public class UserService {
+    private final UserRepository userRepository;
+
+    @Transactional(readOnly = true)
+    public User findById(Long id) {
+        return userRepository.findById(id).get();
+    }
+}
