--- src/main/java/com/unsolvedwa/unsolvedwa/problem/Problem.java
@@ -1,28 +1,0 @@
-package com.unsolvedwa.unsolvedwa.problem;
-
-import com.unsolvedwa.unsolvedwa.BaseTimeEntity;
-import com.unsolvedwa.unsolvedwa.team.ProblemTeam;
-import lombok.Getter;
-
-import javax.persistence.*;
-import java.util.List;
-
-@Entity
-@Getter
-public class Problem extends BaseTimeEntity {
-    @Id @GeneratedValue
-    @Column(name = "problem_id")
-    private Long id;
-
-    private String title;
-
-    private Long tier;
-
-    @OneToMany(mappedBy = "problem", cascade = CascadeType.ALL)
-    private List<ProblemTeam> problemTeams;
-
-    public void setBasicInfo(String title, Long tier) {
-        this.title = title;
-        this.tier = tier;
-    }
-}
