--- src/main/java/com/unsolvedwa/unsolvedwa/problem/ProblemRepository.java
@@ -1,7 +1,0 @@
-package com.unsolvedwa.unsolvedwa.problem;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProblemRepository extends JpaRepository<Problem, Long>, ProblemRepositoryCustom {
-
-}
