--- src/main/java/com/unsolvedwa/unsolvedwa/problem/ProblemRepositoryCustom.java
@@ -1,6 +1,0 @@
-package com.unsolvedwa.unsolvedwa.problem;
-
-import java.util.List;
-
-public interface ProblemRepositoryCustom {
-}
