--- src/main/java/com/unsolvedwa/unsolvedwa/team/TeamRepository.java
@@ -1,7 +1,0 @@
-package com.unsolvedwa.unsolvedwa.team;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TeamRepository extends JpaRepository<Team, Long>, TeamRepositoryCustom {
-
-}
