--- src/main/java/com/unsolvedwa/unsolvedwa/team/TeamRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.unsolvedwa.unsolvedwa.team;
-
-public interface TeamRepositoryCustom {
-}
