--- src/main/java/com/unsolvedwa/unsolvedwa/team/TeamRepositoryImpl.java
@@ -1,11 +1,0 @@
-package com.unsolvedwa.unsolvedwa.team;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public class TeamRepositoryImpl implements TeamRepositoryCustom{
-
-    private JPAQueryFactory queryFactory;
-
-}
