--- src/main/java/com/unsolvedwa/unsolvedwa/user/UserRepository.java
@@ -1,6 +1,0 @@
-package com.unsolvedwa.unsolvedwa.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long>, UserRepositoryCustom{
-}
