--- src/main/java/com/unsolvedwa/unsolvedwa/user/UserRepositoryCustom.java
@@ -1,7 +1,0 @@
-package com.unsolvedwa.unsolvedwa.user;
-
-import com.unsolvedwa.unsolvedwa.team.ProblemTeam;
-
-public interface UserRepositoryCustom {
-    ProblemTeam[] solvingProblem(Long user_id, Long boj_id);
-}
